refactor(shared-architecture): 🔧 rename components and update import paths for better organization
