+++ frontend/src/modules/cart/pages/cartOrdersList.js
@@ -1,8 +1,8 @@
-import Navbar from "../../../shared/components/layout/header/Navbar";
+import Navbar from "../../../shared/components/organisms/Navbar";
 import React, {useEffect, useState} from "react";
-import GoBackButton from "../../../shared/components/ui/Buttons/goBack";
+import GoBackButton from "../../../shared/components/molecules/GoBackButton";
 import {useGetCart} from "../hooks/useGetCart";
-import Loader from "../../../shared/components/ui/Loaders/Loader";
+import Loader from "../../../shared/components/molecules/Loader";
 import {IoMdCart} from "react-icons/io";
 import WhatsAppButton from "../components/WhatsAppButton";
 import {RiDeleteBin5Fill} from "react-icons/ri";
@@ -10,7 +10,7 @@ import useDeleteProductCart from "../hooks/useDeleteProductCart";
 import HorizontalNavBar from "../components/HorizontalNavBar";
 import useFilteredCarts from "../hooks/useFilteredCarts";
 import {handleWhatsAppClick} from "../utils/whatsappUtils";
-import DeleteModal from "../../../shared/components/ui/Modals/DeleteModal";
+import DeleteModal from "../../../shared/components/molecules/DeleteModal";
 import {AnimatePresence, motion} from "framer-motion";
 import updateCartQuantity from "../utils/updateCartQuantity";
 import calculateTotals from "../utils/calculateTotals";
