+++ frontend/src/shared/components/atoms/ButtonCancel.js
@@ -1,5 +1,5 @@
 
-const CancelButton = ({route}) => {
+const ButtonCancel = ({route}) => {
     return (
         <a href={route}
            className="w-1/2 text-center bg-red-500 text-white py-2 rounded-lg hover:bg-red-600 transition">
@@ -8,4 +8,4 @@ const CancelButton = ({route}) => {
     );
 };
 
-export default CancelButton;
+export default ButtonCancel;
