+++ frontend/src/shared/components/atoms/NotificationIcon.js
