+++ frontend/src/shared/components/molecules/DeleteModal.js
