+++ generate_mock_coverage.py
@@ -47,7 +47,7 @@ def create_mock_coverage_xml():
         "frontend/src/core/constants/baseUrls.js",
         "frontend/src/modules/dashboard/components/animations/BoxOfCardsAnimation.js",
         "frontend/src/modules/cart/utils/calculateTotals.js",
-        "frontend/src/shared/components/buttons/CancelButton.js",
+        "frontend/src/shared/components/buttons/ButtonCancel.js",
         "frontend/src/modules/startup/components/card/CardStartup.js",
         "frontend/src/modules/cart/pages/cartOrdersList.js",
         "frontend/src/core/constants/colors/categoryColors.js",
@@ -94,7 +94,7 @@ def create_mock_coverage_xml():
         "frontend/src/modules/startup/services/startupGet.js",
         "frontend/src/modules/startup/context/StartupProvider.js",
         "frontend/src/modules/startup/services/startupService.js",
-        "frontend/src/shared/components/buttons/SubmitButton.js",
+        "frontend/src/shared/components/buttons/ButtonSubmit.js",
         "frontend/src/modules/cart/utils/updateCartQuantity.js",
         "frontend/src/modules/category/services/updateCategoryService.js",
         "frontend/src/modules/products/services/updateProductService.js",
@@ -231,7 +231,7 @@ def create_mock_coverage_xml():
             "frontend/src/core/constants/baseUrls.js": 4,
             "frontend/src/modules/dashboard/components/animations/BoxOfCardsAnimation.js": 2,
             "frontend/src/modules/cart/utils/calculateTotals.js": 2,
-            "frontend/src/shared/components/buttons/CancelButton.js": 2,
+            "frontend/src/shared/components/buttons/ButtonCancel.js": 2,
             "frontend/src/modules/startup/components/card/CardStartup.js": 17,
             "frontend/src/modules/cart/pages/cartOrdersList.js": 24,
             "frontend/src/core/constants/colors/categoryColors.js": 1,
@@ -278,7 +278,7 @@ def create_mock_coverage_xml():
             "frontend/src/modules/startup/services/startupGet.js": 1,
             "frontend/src/modules/startup/context/StartupProvider.js": 4,
             "frontend/src/modules/startup/services/startupService.js": 2,
-            "frontend/src/shared/components/buttons/SubmitButton.js": 2,
+            "frontend/src/shared/components/buttons/ButtonSubmit.js": 2,
             "frontend/src/modules/cart/utils/updateCartQuantity.js": 2,
             "frontend/src/modules/category/services/updateCategoryService.js": 5,
             "frontend/src/modules/products/services/updateProductService.js": 4,
